Refactored according to Igor's comments
