+++ js/main.js
@@ -1,10 +1,10 @@
 // Main
 
 import './validation.js'
-import {setFormActivity, submitForm, form} from'./form.js'
+import {setFormActivity} from'./form.js'
 import {setFilter} from './filter.js'
 import {getData} from './backend.js'
-import {appendCardsToMap, initEmptyMap, resetMainMarker} from './map.js'
+import {appendCardsToMap, initEmptyMap} from './map.js'
 import {showMessage} from './utils.js'
 
 const onSuccessRecieveData = (data) =>  {
@@ -17,15 +17,6 @@ const onErrorRecieveData = () => {
   showMessage('notification', 'error', 'Произошла ошибка')
 }
 
-const onSuccessSubmitData = () =>  {
-  showMessage('success')
-  form.reset()
-  resetMainMarker()
-}
-
-const onErrorSubmitData = () => showMessage('error')
-
 setFilter('map__filters--disabled', 'add')            // отключаем фильтры
 setFormActivity('ad-form--disabled', 'add')           // отключаем элементы формы
 getData(onSuccessRecieveData, onErrorRecieveData)     // получаем данные или инициализация карты с маркером
-submitForm(onSuccessSubmitData, onErrorSubmitData)    // передача данных на сервер
